Migrate project scripts down to boilerdata package level
