+++ .vscode/settings.json
@@ -4,10 +4,9 @@
 
   //! Schema
   "yaml.schemas": {
-    "project/schema/axes_schema.json": "axes.yaml",
-    "project/schema/config_old_schema.json": "config_old.yaml",
-    "project/schema/project_schema.json": "project*.yaml",
-    "project/schema/trials_schema.json": "trials*.yaml"
+    "src/boilerdata/schema/axes_schema.json": "axes.yaml",
+    "src/boilerdata/schema/project_schema.json": "project*.yaml",
+    "src/boilerdata/schema/trials_schema.json": "trials*.yaml"
   },
 
   //! Pylance
