+++ src/boilerdata/config/project.yaml
@@ -1,7 +1,7 @@
 dirs:
   base: "G:/My Drive/Blake/School/Grad/Projects/18.09 Nucleate Pool Boiling/Data"
-  config: "~/Code/mine/boilerdata/project/config"
-  project_schema: "~/Code/mine/boilerdata/project/schema"
+  config: "~/Code/mine/boilerdata/src/boilerdata/config"
+  project_schema: "~/Code/mine/boilerdata/src/boilerdata/schema"
   trials: "Boiling Curves"
   per_trial: "data"
   results: "Results"
