+++ src/boilerdata/config/trials.yaml
