+++ src/boilerdata/schema/axes_schema.json
