+++ src/boilerdata/schema/trials_schema.json
