--- src/boilerdata/typing.py
@@ -1,8 +1,0 @@
-from os import PathLike
-from typing import Any
-
-import numpy as np
-
-StrPath = str | PathLike[str]
-NpNDArray = np.ndarray[Any, Any]
-NpFloating = np.floating[Any]
